Create separate module for StateHistory benchmarks
